Implemented all of the abstract classes in i_map_data. All functions are complete as well, albeit untested. We need to devise some tests for these functions.
Some changes to i_map_data were necessary.
The majority of the changes were changing
the constructors to const, since after we fill
the data, it should not change again in the
program.

The only major change was in the IBuilding
class. The methods getNodeIds and getNodeIdsByCategory
both needed access to the INode List stored in
MapData. This is impossible normally as the
IBuildings class is disconnected from the MapData
class. Therefore, as the frontend will have access
to the MapData class, I have added the INode List
as a parameter to both of those functions. This easily
gives the function access to the Nodes list, and it
doesnt cost any more storage as lists are passed by
reference.